+++ R/sensobol-package.R
@@ -1,4 +1,4 @@
-#' It allows to rapidly compute, bootstrap and plot up to third-order Sobol'-based sensitivity
+#' It allows to rapidly compute, bootstrap and plot up to fourth-order Sobol'-based sensitivity
 #' indices using several state-of-the-art first and total-order estimators. Sobol' indices
 #' can be computed either for models that yield a scalar as a model output or for systems of
 #' differential equations. The package also provides a suit of benchmark tests functions